% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork_helpers.R
\name{add_steering_axis}
\alias{add_steering_axis}
\title{Add steering axis, adjusted fork and rake triangle}
\usage{
add_steering_axis(frame_dims, fork_rake = 45)
}
\arguments{
\item{frame_dims}{A 6 x 3 tibble with frame dimensions from the front
and rear triangle, returned by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}}.}

\item{fork_rake}{The fork rake in millimeters}
}
\value{
A 10 x 3 tibble.
}
\description{
The adjusted fork and the actual fork are two different
things. The former is the hypotenuse of a right triangle
whose legs are the steering axis and the rake line. The
latter is the hypotenuse of a right triangle whose short
leg is the rake line and whose long leg runs along the
steering axis, but is shorter. The difference is the
head tube extension.
}
\seealso{
\code{\link[=wrap_frame_dims]{wrap_frame_dims()}}
}
