% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_the_frame.R
\name{add_the_ett_to_the_drawing}
\alias{add_the_ett_to_the_drawing}
\title{Add the effective top tube to a frame drawing}
\usage{
add_the_ett_to_the_drawing(
  x,
  df,
  wheel_diameter = 622,
  c = I("black"),
  alpha_factor = 1
)
}
\arguments{
\item{x}{An existing ggplot object. At a minimum, it's a scatter plot of
the rear dropout center. But it can be a bicycle frame of a previous design
already drawn, onto which you can now overlay a new design using the dims
in \code{df} for easy comparison.}

\item{df}{A frame design tibble, such as the minimal one drawn by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}}
or the one augmented by \code{\link[=find_ht_extension_and_add_true_fork]{find_ht_extension_and_add_true_fork()}}.}

\item{wheel_diameter}{Wheel diameter in millimeters; 622 is the bead seat diameter
of a 700c wheel. To see the effect of the tire width, add 2 x tire width in millimeters
-- i.e. specify \code{wheel_diameter = 622 + 2 * 32} for a 32 mm tire.}

\item{c}{The color of the line segments.}

\item{alpha_factor}{The transparency of the line segments, \code{alpha}.}
}
\value{
A ggplot object.
}
\description{
Useful for slant top tube designs. For horizontal
top tube designs there's nothing to add and you
can check in \code{df}: the elements of the \code{ett_triangle}
vector will be all 0.
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
