% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_the_frame.R
\name{add_the_wheels}
\alias{add_the_wheels}
\title{Add the wheels to a frame drawing}
\usage{
add_the_wheels(
  x,
  df,
  bb_shell_diameter = 34.8,
  wheel_diameter = 622,
  c = I("black"),
  alpha_factor = 1
)
}
\arguments{
\item{x}{An existing ggplot object. At a minimum, it's a scatter plot of
the rear dropout center. But it can be a bicycle frame of a previous design
already drawn, onto which you can now overlay a new design using the dims
in \code{df} for easy comparison.}

\item{df}{A frame design tibble, such as the minimal one drawn by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}}
or the one augmented by \code{\link[=find_ht_extension_and_add_true_fork]{find_ht_extension_and_add_true_fork()}}.}

\item{bb_shell_diameter}{Outer diameter of the BB shell in millimeters.}

\item{wheel_diameter}{Wheel diameter in millimeters; 622 is the bead seat diameter
of a 700c wheel. To see the effect of the tire width, add 2 x tire width in millimeters
-- i.e. specify \code{wheel_diameter = 622 + 2 * 32} for a 32 mm tire.}

\item{c}{The color of the line segments.}

\item{alpha_factor}{The transparency of the line segments, \code{alpha}.}
}
\value{
A ggplot object.
}
\description{
Adding the wheels to the frame plot may help you get an idea of the overall
proportions. Other than that, it's not a very useful thing to do.
\emph{Warning}: this requires a frame design tibble \code{df} that has the columns
added by \code{\link[=find_ht_extension_and_add_true_fork]{find_ht_extension_and_add_true_fork()}}. A
minimal \code{df} tibble as drawn by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}} will
not be sufficient. If all went well, when we draw a fork starting at
the bottom end of the HT extension, with a given rake and cta length,
it will end up exactly at the top end of the rake line.
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
