% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_dimensions.R
\name{add_tube_extension}
\alias{add_tube_extension}
\title{Add extension to top of the head or seat tube}
\usage{
add_tube_extension(frame_dims, tube = "ht_triangle", t_extension = 20)
}
\arguments{
\item{frame_dims}{A tibble with frame dimensions from the front
and rear triangle, returned by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}}.}

\item{tube}{The name of the tube to be extended: \code{ht_triangle}
for the head tube or \code{st_triangle} for the seat tube.}

\item{t_extension}{The length of the tube extension in millimeters}
}
\value{
A tibble.
}
\description{
The frame reach is the difference between the x coordinate
of the top of the head tube and the x coordinate of the
bottom bracket. This is measured at the true top of the head
tube, which is somewhere above the point where the top tube
and the head tube meet. The true length of the head tube is
a derived dimension, but the distance between the top of the
head tube and the point where the top tube and the head tube
meet, set here as \code{t_extension}, can be an arbitrary input if
you have a welded frame. In a lugged frame the top of the head
tube is the same as the top of the HT-TT lug ring.
}
\details{
The same goes for the top of the seat tube extension. It too
is arbitrary for a welded frame, and the trigonometric
calculations for deriving its vertical and horizontal projections
are identical to those of the head tube extension so we might
as well use the same function for getting at them.
}
\seealso{
\code{\link[=get_front_triangle_dims]{get_front_triangle_dims()}}
}
