% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork_helpers.R
\name{calculate_fork_trail}
\alias{calculate_fork_trail}
\title{Calculate fork trail}
\usage{
calculate_fork_trail(df, wheel_diameter = 622)
}
\arguments{
\item{df}{A tibble with frame dimensions returned by \code{\link[=find_ht_extension_and_add_true_fork]{find_ht_extension_and_add_true_fork()}}.}

\item{wheel_diameter}{Wheel diameter in millimeters; 622 is the bead seat diameter
of a 700c wheel. To see the effect of the tire width, add 2 x tire width in millimeters}
}
\value{
A scalar equal to the fork trail in millimeters
}
\description{
Given a tibble with all the frame dimensions you get
after you run both \code{\link[=wrap_frame_dims]{wrap_frame_dims()}} and
\code{\link[=find_ht_extension_and_add_true_fork]{find_ht_extension_and_add_true_fork()}} you
can now calculate the fork trail.
}
