% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_the_frame.R
\name{draw_the_bicycle}
\alias{draw_the_bicycle}
\title{Draw a fresh frame onto a blank sheet}
\usage{
draw_the_bicycle(
  df,
  wheel_diameter = 622,
  i = 350,
  c = I("black"),
  alpha_factor = 1
)
}
\arguments{
\item{df}{A frame design tibble, such as drawn by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}}.}

\item{wheel_diameter}{Wheel diameter in millimeters; 622 is the bead seat diameter
of a 700c wheel. To see the effect of the tire width, add 2 x tire width in millimeters
-- i.e. specify \code{wheel_diameter = 622 + 2 * 32} for a 32 mm tire.}

\item{i}{The offset from x = 0, so the frame fits nicely in the first quadrant.}

\item{c}{The color of the line segments.}

\item{alpha_factor}{The transparency of the line segments, \code{alpha}.}
}
\value{
A ggplot object.
}
\description{
This wraps \code{\link[=overlay_the_bicycle]{overlay_the_bicycle()}} for convenience.
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
