% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork_helpers.R
\name{find_ht_extension_and_add_true_fork}
\alias{find_ht_extension_and_add_true_fork}
\title{Add the true fork}
\usage{
find_ht_extension_and_add_true_fork(df, fork_rake = 45, fork_cta_length = 390)
}
\arguments{
\item{fork_rake}{The fork rake in millimeters}

\item{fork_cta_length}{The fork crown-to-axle length in millimeters.}
}
\value{
An 11 x 3 tibble.
}
\description{
The fork is specified with a given crown-to-axle length and
rake. These determine ride characteristics, wheel base length,
and toe overlap. Depending on how long the fork is, the head
tube must extend below the point where it meets the down tube
so we will have to add that bit to the frame dimensions.
}
\seealso{
\code{\link[=wrap_frame_dims]{wrap_frame_dims()}}
}
