% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{get_all_the_metrics}
\alias{get_all_the_metrics}
\title{Get some useful frame metrics}
\usage{
get_all_the_metrics(
  df,
  angle_btw_css = 14,
  wheel_diameter = 622,
  crank_arm_length = 170,
  old_spacing = 130,
  bbs_width = 68
)
}
\arguments{
\item{df}{A frame design tibble, such as the minimal one drawn by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}}
or the one augmented by \code{\link[=find_ht_extension_and_add_true_fork]{find_ht_extension_and_add_true_fork()}}.}

\item{angle_btw_css}{Angle between chain stays, in degrees.}

\item{wheel_diameter}{Wheel diameter in millimeters; 622 is the bead seat diameter
of a 700c wheel. To see the effect of the tire width, add 2 x tire width in millimeters
-- i.e. specify \code{wheel_diameter = 622 + 2 * 32} for a 32 mm tire.}

\item{crank_arm_length}{The crank arm length in millimeters. Typical values are 165, 170 and 175 for adult riders.}

\item{old_spacing}{Over-locknut-dimension (hub width) in millimeters.}

\item{bbs_width}{Bottom bracket shell width. It's standard -- 68 mm is typical for road bikes.}
}
\value{
A named vector
}
\description{
Once you have a frame design \code{df} you can use this to derive
some metrics that will help gauge the ride characteristics.
}
