% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bb_drop.R
\name{get_bb_drop}
\alias{get_bb_drop}
\title{Get the bottom bracket drop}
\usage{
get_bb_drop(cs_st_angle, st_angle, cs_length, angle_btw_css)
}
\arguments{
\item{cs_st_angle}{The angle between the chain stay (cs) and seat tube (st) in degrees.}

\item{st_angle}{The angle between the seat tube and the horizontal.}

\item{cs_length}{Chain stay length in millimeters.}

\item{angle_btw_css}{Angle between chain stays, in degrees.}
}
\value{
A scalar equal to the bottom bracket drop in millimeters.
}
\description{
Get the bottom bracket drop
}
\examples{
get_bb_drop(60, 74, 450, 14)
}
