% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bb_tt_diagonal.R
\name{get_bb_tt_diagonal}
\alias{get_bb_tt_diagonal}
\title{Get the length of the front triangle diagonal}
\usage{
get_bb_tt_diagonal(st_length, tt_length, st_angle, tt_angle)
}
\arguments{
\item{st_length}{The length of the seat tube ST in millimeters.}

\item{tt_length}{The length of the top tube TT in millimeters.}

\item{st_angle}{The angle between the seat tube and the horizontal.}

\item{tt_angle}{The angle between the top tube and the horizontal.}
}
\value{
A scalar equal to the length of the diagonal in millimeters.
}
\description{
The front triangle is a misnomer, because it's a quadrilateral made up
by the seat tube ST, top tube TT, head tube HT and down tube DT. The
diagonal we want is the second one, going from the bottom bracket BB
to the point where the top tube meets the head tube.
}
\examples{
get_bb_tt_diagonal(500, 530, 73, 71)
}
