% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_helpers.R
\name{get_cs_length}
\alias{get_cs_length}
\title{Get the chain stay length}
\usage{
get_cs_length(old_spacing, angle_btw_css, bbsc_offset)
}
\arguments{
\item{old_spacing}{Over-locknut-dimension (hub width) in millimeters.}

\item{angle_btw_css}{Angle between chain stays, in degrees.}

\item{bbsc_offset}{The bottom bracket shell offset in millimeters.}
}
\value{
A scalar equal to the chain stay length in millimeters.
}
\description{
In the plane that is perpendicular onto the bike, the chain
stay is part of a trapezoid that is as wide as half the OLD
spacing at the dropout and half the bottom bracket shell
center offset at the bottom bracket. The OLD spacing is standard
(130 mm for road bikes, 135 mm for old-school mountain bikes,
142 mm for newer mountain bikes, etc.). So the OLD spacing is
an input in frame design, not a derived dimension. If you have
a bottom bracket shell with a given angle between chain stays
(\code{angle_btw_css}) and center offset (\code{bbsc_offset}) then you
must use this function to determine the chain stay length that
will produce the OLD intended for your frame.
}
\examples{
get_cs_length(130, 10, 10)
}
