% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_dimensions.R
\name{get_front_triangle_dims}
\alias{get_front_triangle_dims}
\title{Get front triangle dimensions}
\usage{
get_front_triangle_dims(
  st_length = 500,
  tt_length = 500,
  st_angle = 71,
  tt_angle = 0,
  ht_angle = 71,
  st_dt_angle = 60
)
}
\arguments{
\item{st_length}{The length of the seat tube ST in millimeters.}

\item{tt_length}{The length of the top tube TT in millimeters.}

\item{st_angle}{The angle between the seat tube and the horizontal.}

\item{tt_angle}{The angle between the top tube and the horizontal.}

\item{ht_angle}{Head tube angle with the horizontal, in degrees.}

\item{st_dt_angle}{Angle between the seat tube and the down tube, in degrees.}
}
\value{
A 3 x 5 tibble.
}
\description{
The front triangle is a misnomer, because it's a quadrilateral made up
by four tubes: the seat tube ST, top tube TT, head tube HT and down tube DT.
Between these tubes there are four angles, and they must add up to 360 degrees.
A bike frame designer aims for a top tube length (\code{tt_length}) and
seat tube length (\code{st_length}) that fit a rider of a given height and
torso length, and also for a seat tube angle (\code{st_angle}) head
tube angle (\code{ht_angle}) and top tube angle (\code{tt_angle}) that give
the bike its desired riding characteristics.
}
\details{
The angle between the seat tube and the downtube (\code{st_dt_angle}) is
pre-determined in lugged frame construction by how the bottom bracket
shell was cast. With these six dimensions in hand we can get the full
set of dimensions that describe the front triangle. This function
returns them as a tibble of five columns and three rows. Each tube is
described as a right triangle made up of the tube length as the hypotenuse
and the tube's horizontal and vertical projections as the legs. The elements
on each row are named accordingly. Why five columns for four tubes?

The fifth column, ett_triangle, helps calculate the effective top tube length
when the top tube is slanted. For slanted top tubes the effective length is
the horizontal distance between the front of the top tube and an imaginary
extension of the seat tube equal to the vertical projection of the top tube
times the sine of the seat tube angle in radians. In other words, this
horizontal distance is equal to the horizontal projection of the top tube
plus the vertical projection of the top tube times the tangent of the seat
tube angle in radians. The second term of this addition is the third
element of the ett_triangle vector.
}
\examples{
get_front_triangle_dims()
}
