% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_helpers.R
\name{get_old_spacing}
\alias{get_old_spacing}
\title{Get the over-locknuts-dimension (OLD) spacing}
\usage{
get_old_spacing(cs_length, angle_btw_css, bbsc_offset)
}
\arguments{
\item{cs_length}{Chain stay length in millimeters.}

\item{angle_btw_css}{Angle between chain stays, in degrees.}

\item{bbsc_offset}{The bottom bracket shell offset in millimeters.}
}
\value{
A scalar equal to the OLD spacing in millimeters.
}
\description{
In the plane that is perpendicular onto the bike, the chain
stay is part of a trapezoid that is as wide as half the OLD
spacing at the dropout and half the bottom bracket shell
center offset at the bottom bracket. The OLD spacing is standard
(130 mm for road bikes, 135 mm for old-school mountain bikes,
142 mm for newer mountain bikes, etc.). So the OLD spacing is
an input in frame design, not a derived dimension. This function
is only useful as a sanity check, to make sure that you hit the
right OLD number with a given set of values for its arguments.
}
\examples{
get_old_spacing(450, 10, 10)
}
