% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_dimensions.R
\name{get_rear_triangle_dims}
\alias{get_rear_triangle_dims}
\title{Get rear triangle dimensions}
\usage{
get_rear_triangle_dims(
  st_length = 500,
  cs_length = 450,
  st_angle = 71,
  cs_st_angle = 61,
  angle_btw_css = 14
)
}
\arguments{
\item{st_length}{The length of the seat tube ST in millimeters.}

\item{cs_length}{Chain stay length in millimeters.}

\item{st_angle}{The angle between the seat tube and the horizontal.}

\item{cs_st_angle}{Angle between chain stay and seat tube, in degrees.}

\item{angle_btw_css}{Angle between chain stays, in degrees.}
}
\value{
A 3 x 3 tibble.
}
\description{
The rear triangle is made up of the projection in the horizontal plane
of the chain stay, the seat stay, and the seat tube. The seat tube is
the same as its projection in the horizontal plane, but the other two
tubes stick out at an angle, ans they must accommodate the width of
the rear wheel hub, so there's some trigonometry involved.
}
\details{
A bike frame designer aims for a seat tube length (\code{st_length}) that
fits a rider of a given height, and also for a seat tube angle (\code{st_angle})
and chain stay length (\code{cs_length}) that give the bike its desired riding
characteristics, though \code{cs_length} is subject to some other constraints
described in the documentation for \code{\link[=get_cs_length]{get_cs_length()}}.

The angle between the chain stay and the seat tube (\code{cs_st_angle}) is
pre-determined in lugged frame construction by how the bottom bracket
shell was cast.

This function returns the dimensions of the rear triangle using the
convention described in the documentation for \code{\link[=get_front_triangle_dims]{get_front_triangle_dims()}}.
}
\examples{
get_rear_triangle_dims()
}
\seealso{
\code{\link[=get_cs_length]{get_cs_length()}}

\code{\link[=get_front_triangle_dims]{get_front_triangle_dims()}}
}
