% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{get_st_to_rear_tire_clearance}
\alias{get_st_to_rear_tire_clearance}
\title{Get seat tube to rear tire clearance}
\usage{
get_st_to_rear_tire_clearance(df, wheel_diameter = 622)
}
\arguments{
\item{df}{A tibble with frame dimensions from the front
and rear triangle, returned by \code{\link[=wrap_frame_dims]{wrap_frame_dims()}} or
at a minimum the portion of it that comes from \code{\link[=get_rear_triangle_dims]{get_rear_triangle_dims()}}}

\item{wheel_diameter}{Wheel diameter in millimeters; 622 is the bead seat diameter
of a 700c wheel. To see the effect of the tire width, add 2 x tire width in millimeters
-- i.e. specify \code{wheel_diameter = 622 + 2 * 32} for a 32 mm tire.}
}
\value{
A scalar, the seat tube to rear tire clearance in millimeters
}
\description{
Frame designs with shorter chain stays are for racing; longer
chain stays are for comfort. You can tell them apart because
the latter have a larger gap between the rear tire and the
seat tube. This function returns that gap, in millimeters,
for a given design tibble \code{df} and wheel diameter.
}
