% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_dimensions.R
\name{wrap_frame_dims}
\alias{wrap_frame_dims}
\title{Get the dimensions of the bicycle frame}
\usage{
wrap_frame_dims(
  st_length = 500,
  tt_length = 500,
  st_angle = 71,
  tt_angle = 0,
  ht_angle = 71,
  st_dt_angle = 60,
  cs_length = 450,
  cs_st_angle = 59,
  angle_btw_css = 14
)
}
\arguments{
\item{st_length}{The length of the seat tube ST in millimeters.}

\item{tt_length}{The length of the top tube TT in millimeters.}

\item{st_angle}{The angle between the seat tube and the horizontal.}

\item{tt_angle}{The angle between the top tube and the horizontal.}

\item{ht_angle}{Head tube angle with the horizontal, in degrees.}

\item{st_dt_angle}{Angle between the seat tube and the down tube, in degrees.}

\item{cs_length}{Chain stay length in millimeters.}

\item{cs_st_angle}{Angle between chain stay and seat tube, in degrees.}

\item{angle_btw_css}{Angle between chain stays, in degrees.}
}
\value{
A 6 x 3 tibble.
}
\description{
This is a convenience wrapper that puts together the
output of \code{\link[=get_front_triangle_dims]{get_front_triangle_dims()}} and
\code{\link[=get_rear_triangle_dims]{get_rear_triangle_dims()}}.
}
\examples{
wrap_frame_dims()
}
\seealso{
\code{\link[=get_front_triangle_dims]{get_front_triangle_dims()}}

\code{\link[=get_rear_triangle_dims]{get_rear_triangle_dims()}}
}
